move splitkey method to SplitKeyUtils
